rabix-bindings: minor refactor
