+++ rabix-bindings/src/test/resources/org/rabix/bindings/protocol/helper/grep-wf-loop.cwl.json
