+++ rabix-bindings/src/test/resources/org/rabix/bindings/protocol/helper/grep-wf.cwl.json
